group4-codes-1/pipeline/requirements.txt
boto3==1.26.142
botocore==1.29.142
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
fonttools==4.39.4
idna==3.4
jmespath==1.0.1
joblib==1.2.0
kaggle==1.5.13
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0
requests==2.31.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
text-unidecode==1.3
threadpoolctl==3.1.0
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.6.2
tzdata==2023.3
urllib3==1.26.16
